+++ src/test/java/com/braintreegateway/integrationtest/WebhookNotificationIT.java
@@ -2,11 +2,13 @@ package com.braintreegateway.integrationtest;
 
 import com.braintreegateway.BraintreeGateway;
 import com.braintreegateway.Environment;
+import com.braintreegateway.MerchantAccount;
 import com.braintreegateway.WebhookNotification;
 import com.braintreegateway.exceptions.InvalidSignatureException;
 import com.braintreegateway.testhelpers.TestHelper;
 import com.braintreegateway.util.NodeWrapper;
 import com.braintreegateway.util.NodeWrapperFactory;
+import com.braintreegateway.ValidationErrorCode;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -25,7 +27,7 @@ public class WebhookNotificationIT {
 
     @Test
     public void createNotificationWithUnrecognizedKind() {
-        String xml = "<notification><kind>" + "bad_kind" + "</kind></notification>";
+        String xml = "<notification><kind>" + "bad_kind" + "</kind> <subject> </subject> </notification>";
         NodeWrapper node = NodeWrapperFactory.instance.create(xml);
 
         WebhookNotification notification = new WebhookNotification(node);
@@ -39,7 +41,7 @@ public class WebhookNotificationIT {
     }
 
     @Test
-    public void createsSampleNotification() {
+    public void createsSampleSubscriptionNotification() {
         HashMap<String, String> sampleNotification = this.gateway.webhookTesting().sampleNotification(WebhookNotification.Kind.SUBSCRIPTION_WENT_PAST_DUE, "my_id");
 
         WebhookNotification notification = this.gateway.webhookNotification().parse(sampleNotification.get("signature"), sampleNotification.get("payload"));
@@ -47,6 +49,42 @@ public class WebhookNotificationIT {
         assertEquals(WebhookNotification.Kind.SUBSCRIPTION_WENT_PAST_DUE, notification.getKind());
         assertEquals("my_id", notification.getSubscription().getId());
         TestHelper.assertDatesEqual(Calendar.getInstance(), notification.getTimestamp());
+    }
+
+    @Test
+    public void createsSampleMerchantAccountApprovedNotification() {
+        HashMap<String, String> sampleNotification = this.gateway.webhookTesting().sampleNotification(WebhookNotification.Kind.SUB_MERCHANT_ACCOUNT_APPROVED, "my_id");
+
+        WebhookNotification notification = this.gateway.webhookNotification().parse(sampleNotification.get("signature"), sampleNotification.get("payload"));
+
+        assertEquals(WebhookNotification.Kind.SUB_MERCHANT_ACCOUNT_APPROVED, notification.getKind());
+        assertEquals("my_id", notification.getMerchantAccount().getId());
+        assertEquals(MerchantAccount.Status.ACTIVE, notification.getMerchantAccount().getStatus());
+        TestHelper.assertDatesEqual(Calendar.getInstance(), notification.getTimestamp());
+    }
+
+    @Test
+    public void createsSampleMerchantAccountDeclinedNotification() {
+        HashMap<String, String> sampleNotification = this.gateway.webhookTesting().sampleNotification(WebhookNotification.Kind.SUB_MERCHANT_ACCOUNT_DECLINED, "my_id");
+
+        WebhookNotification notification = this.gateway.webhookNotification().parse(sampleNotification.get("signature"), sampleNotification.get("payload"));
+
+        assertEquals(WebhookNotification.Kind.SUB_MERCHANT_ACCOUNT_DECLINED, notification.getKind());
+        assertEquals("my_id", notification.getMerchantAccount().getId());
+        assertEquals(MerchantAccount.Status.SUSPENDED, notification.getMerchantAccount().getStatus());
+        TestHelper.assertDatesEqual(Calendar.getInstance(), notification.getTimestamp());
+    }
+
+    @Test
+    public void createsSampleMerchantAccountDeclinedNotificationWithErrorCodes() {
+        HashMap<String, String> sampleNotification = this.gateway.webhookTesting().sampleNotification(WebhookNotification.Kind.SUB_MERCHANT_ACCOUNT_DECLINED, "my_id");
+
+        WebhookNotification notification = this.gateway.webhookNotification().parse(sampleNotification.get("signature"), sampleNotification.get("payload"));
+
+        assertEquals(WebhookNotification.Kind.SUB_MERCHANT_ACCOUNT_DECLINED, notification.getKind());
+        assertEquals("my_id", notification.getMerchantAccount().getId());
+        TestHelper.assertDatesEqual(Calendar.getInstance(), notification.getTimestamp());
+        assertEquals(ValidationErrorCode.MERCHANT_ACCOUNT_APPLICANT_DETAILS_DECLINED_OFAC, notification.getErrors().forObject("merchantAccount").onField("base").get(0).getCode());
     }
 
     @Test(expected = InvalidSignatureException.class)
@@ -61,5 +99,18 @@ public class WebhookNotificationIT {
         HashMap<String, String> sampleNotification = this.gateway.webhookTesting().sampleNotification(WebhookNotification.Kind.SUBSCRIPTION_WENT_PAST_DUE, "my_id");
 
         this.gateway.webhookNotification().parse("uknown_public_key|signature", sampleNotification.get("payload"));
+    }
+
+    @Test
+    public void createsSampleTransactionDisbursedNotification() {
+        HashMap<String, String> sampleNotification = this.gateway.webhookTesting().sampleNotification(WebhookNotification.Kind.TRANSACTION_DISBURSED, "my_id");
+
+        WebhookNotification notification = this.gateway.webhookNotification().parse(sampleNotification.get("signature"), sampleNotification.get("payload"));
+
+        assertEquals(WebhookNotification.Kind.TRANSACTION_DISBURSED, notification.getKind());
+        assertEquals("my_id", notification.getTransaction().getId());
+        assertEquals(2013, notification.getTransaction().getDisbursementDetails().getDisbursementDate().get(Calendar.YEAR));
+        assertEquals(Calendar.JULY, notification.getTransaction().getDisbursementDetails().getDisbursementDate().get(Calendar.MONTH));
+        assertEquals(9, notification.getTransaction().getDisbursementDetails().getDisbursementDate().get(Calendar.DAY_OF_MONTH));
     }
 }