+++ src/test/java/com/braintreegateway/testhelpers/MerchantAccountTestConstants.java
@@ -1,7 +1,10 @@
-package com.braintreegateway.integrationtest;
+package com.braintreegateway.testhelpers;
 
-public class MerchantAccount {
+public interface MerchantAccountTestConstants {
     public static final String DEFAULT_MERCHANT_ACCOUNT_ID = "sandbox_credit_card";
     public static final String NON_DEFAULT_MERCHANT_ACCOUNT_ID = "sandbox_credit_card_non_default";
     public static final String NON_DEFAULT_SUB_MERCHANT_ACCOUNT_ID = "sandbox_sub_merchant_account";
+
+    public static final String SANDBOX_MASTER_MERCHANT_ACCOUNT = "sandbox_master_merchant_account";
+
 }